requirements.txt
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.0
supabase>=2.3.0
python-dotenv>=1.0.0
moviepy>=1.0.3
Pillow>=10.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
httpx>=0.25.0
aiohttp>=3.9.0
schedule>=1.2.0
fastapi>=0.104.0
uvicorn>=0.24.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
